Create code
This code takes the text and a list of search words as input into the search_words function. It splits the text into individual words and checks if each word is present in the list of search words. The found words are added to the found_words list, which is returned as the result.

In this example, the result will be ['example', 'text', 'search'] as all the search words are found in the text.

Please note that this code does not consider word case. It treats "example" and "Example" as different words. If you need a case-insensitive search, you can convert all the words to lowercase using the .lower() method before comparison.